+++ packages/reducers/src/makeReducer.js
@@ -1,4 +1,3 @@
-import invariant from 'invariant';
 import {
 	prop,
 	compose,
@@ -19,9 +18,7 @@ const createTypeEqualsPredicate = type => compose(whereEq({ type }), nthArg(1));
 
 const mergeReducers = ([typePredicate, reducer, errorReducer]) => {
 	const newReducer = (state, action) => {
-		if (prop('error', action)) {
-			invariant(errorReducer, `You haven't supplied an error reducer for action ${action.type}`);
-
+		if (prop('error', action) && errorReducer) {
 			return errorReducer(state, action);
 		}